+++ library/Kathu/Scripting/Lua/Global.hs
@@ -9,13 +9,13 @@ import           Linear.V2                 (V2(..))
 import qualified System.Random             as R
 import           Verda.Event.Controls      (CursorMotionState, cursorPosition)
 import           Verda.Logger
+import           Verda.Time
 import           Verda.Util.Containers     (fromJustElseError)
 import           Verda.Util.Types
 import           Verda.Util.Apecs
 
 import           Kathu.Entity.Components
 import           Kathu.Entity.System
-import           Kathu.Entity.Time
 import           Kathu.Graphics.Camera
 import           Kathu.Scripting.ExternalFunctions
 import           Kathu.Scripting.Lua.Types