+++ verda/ChangeLog.md
@@ -2,3 +2,7 @@
 
 ## 0.1.0.0 -- TBD
 Initial merger of verda game engine. New features are as follows:
+* Texture scaling can be set on a per file basis with @nearest or @linear file name suffixes
+* Sound clip and music playing
+Issues to resolve before merging to master:
+* Palettes no longer work, need to implement shader behavior for this